models/sampling/attention_based_sampler.py
"""
Attention-Based Sampling implementation for ProteinFlex.
Implements structure-aware attention routing for protein generation.
"""

import torch
import torch.nn as nn
import torch.nn.functional as F
from typing import Dict, Tuple, Optional

class StructureAwareAttention(nn.Module):
    def __init__(
        self,
        feature_dim: int,
        num_heads: int = 8,
        dropout: float = 0.1
    ):
        """Initialize structure-aware attention."""
        super().__init__()
        self.num_heads = num_heads
        self.head_dim = feature_dim // num_heads
        assert self.head_dim * num_heads == feature_dim, "feature_dim must be divisible by num_heads"

        self.qkv = nn.Linear(feature_dim, 3 * feature_dim)
        self.structure_proj = nn.Linear(feature_dim, feature_dim)
        self.output_proj = nn.Linear(feature_dim, feature_dim)
        self.dropout = nn.Dropout(dropout)

    def forward(
        self,
        x: torch.Tensor,
        structure_bias: Optional[torch.Tensor] = None
    ) -> torch.Tensor:
        """Forward pass with optional structure bias."""
        B, L, D = x.shape
        qkv = self.qkv(x).reshape(B, L, 3, self.num_heads, self.head_dim).permute(2, 0, 3, 1, 4)
        q, k, v = qkv[0], qkv[1], qkv[2]

        # Compute attention scores
        attn = (q @ k.transpose(-2, -1)) * (1.0 / torch.sqrt(torch.tensor(self.head_dim)))

        # Add structure bias if provided
        if structure_bias is not None:
            structure_weights = self.structure_proj(structure_bias)
            structure_weights = structure_weights.view(B, 1, L, L)
            attn = attn + structure_weights

        attn = F.softmax(attn, dim=-1)
        attn = self.dropout(attn)

        # Apply attention to values
        x = (attn @ v).transpose(1, 2).reshape(B, L, D)
        x = self.output_proj(x)

        return x

class AttentionBasedSampler(nn.Module):
    def __init__(
        self,
        feature_dim: int = 768,
        hidden_dim: int = 512,
        num_layers: int = 6,
        num_heads: int = 8,
        dropout: float = 0.1
    ):
        """
        Initialize Attention-Based Sampler.

        Args:
            feature_dim: Dimension of protein features
            hidden_dim: Hidden dimension for feed-forward
            num_layers: Number of transformer layers
            num_heads: Number of attention heads
            dropout: Dropout rate
        """
        super().__init__()
        self.feature_dim = feature_dim
        self.hidden_dim = hidden_dim

        # Structure encoder
        self.structure_encoder = nn.Sequential(
            nn.Linear(feature_dim, hidden_dim),
            nn.LayerNorm(hidden_dim),
            nn.ReLU(),
            nn.Linear(hidden_dim, feature_dim)
        )

        # Transformer layers
        self.layers = nn.ModuleList([
            nn.ModuleDict({
                'attention': StructureAwareAttention(feature_dim, num_heads, dropout),
                'norm1': nn.LayerNorm(feature_dim),
                'ff': nn.Sequential(
                    nn.Linear(feature_dim, hidden_dim),
                    nn.ReLU(),
                    nn.Dropout(dropout),
                    nn.Linear(hidden_dim, feature_dim)
                ),
                'norm2': nn.LayerNorm(feature_dim)
            }) for _ in range(num_layers)
        ])

        # Output projection
        self.output_proj = nn.Linear(feature_dim, feature_dim)

    def forward(
        self,
        x: torch.Tensor,
        structure_info: Optional[torch.Tensor] = None
    ) -> torch.Tensor:
        """
        Forward pass for training.

        Args:
            x: Input protein features [batch_size, seq_len, feature_dim]
            structure_info: Optional structure information

        Returns:
            Processed features
        """
        # Process structure information if provided
        structure_bias = None
        if structure_info is not None:
            structure_bias = self.structure_encoder(structure_info)

        # Apply transformer layers
        for layer in self.layers:
            # Attention with structure bias
            attn_out = layer['attention'](
                layer['norm1'](x),
                structure_bias
            )
            x = x + attn_out

            # Feed-forward
            ff_out = layer['ff'](layer['norm2'](x))
            x = x + ff_out

        return self.output_proj(x)

    def sample(
        self,
        batch_size: int,
        seq_len: int,
        device: torch.device,
        structure_info: Optional[torch.Tensor] = None,
        temperature: float = 1.0
    ) -> torch.Tensor:
        """
        Generate protein features using attention-based sampling.

        Args:
            batch_size: Number of samples to generate
            seq_len: Sequence length
            device: Device to generate on
            structure_info: Optional structure information
            temperature: Sampling temperature

        Returns:
            Generated protein features
        """
        # Initialize from random
        x = torch.randn(batch_size, seq_len, self.feature_dim, device=device)

        # Apply temperature scaling
        x = x * temperature

        # Generate features with structure guidance
        return self.forward(x, structure_info)

    def compute_loss(
        self,
        pred_features: torch.Tensor,
        target_features: torch.Tensor,
        structure_info: Optional[torch.Tensor] = None
    ) -> torch.Tensor:
        """
        Compute training loss.

        Args:
            pred_features: Predicted protein features
            target_features: Target protein features
            structure_info: Optional structure information

        Returns:
            Loss value
        """
        # Feature reconstruction loss
        recon_loss = F.mse_loss(pred_features, target_features)

        # Structure-aware loss if structure info provided
        if structure_info is not None:
            structure_pred = self.structure_encoder(pred_features)
            structure_loss = F.mse_loss(structure_pred, structure_info)
            return recon_loss + structure_loss

        return recon_loss
